已经解构代码user_service，并修改了command中代码使其适配结构后的文件
